Refactor auth composables into nested folders
