+++ pages/src/features/auth/composables/register/useRegisterValidation.ts
