+++ pages/src/features/auth/views/login/LoginView.setup.ts
@@ -1,4 +1,4 @@
-import { useLoginLogic } from '@auth/composables/useLoginLogic';
+import { useLoginLogic } from '@auth/composables/login/useLoginLogic';
 import '@auth/styles/auth.css';
 
 export function useLoginViewSetup() {
