+++ pages/src/features/auth/views/register/RegisterView.setup.ts
@@ -1,4 +1,4 @@
-import { useRegisterLogic } from '@auth/composables/useRegisterLogic';
+import { useRegisterLogic } from '@auth/composables/register/useRegisterLogic';
 import '@auth/styles/auth.css';
 
 export function useRegisterViewSetup() {
